+++ tsdown.config.ts
@@ -21,7 +21,8 @@ export default defineConfig({
   external: [
     'react',
     'react/jsx-runtime',
-    '@deepseek-ai/dsh-client-runtime/client',
+    '@deepseek-ai/cordis',
+    '@deepseek-ai/dsh-client-store',
     '@deepseek-ai/dsh-client-ui-primitives',
   ],
   noExternal: ['zod'],